% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAIC.R
\name{cAIC}
\alias{cAIC}
\title{Calculate Conditional AIC for a fitted Langevin model}
\usage{
cAIC(fit, data, spatialCovs, nSims = 200)
}
\arguments{
\item{fit}{A \code{fitLangevin} object.}

\item{data}{A \code{dataLangevin} object used to fit the model.}

\item{spatialCovs}{A list of named \code{\link[terra]{SpatRaster-class}} objects containing the spatial covariates.}

\item{nSims}{Integer. The number of Rademacher vectors used for Hutchinson's trace estimator (if the exact sparse inverse fails). Default: 200.}
}
\value{
A list containing the calculated \code{cAIC}, effective degrees of freedom (\code{EDF}),
the conditional negative log-likelihood (\code{NLL_cond}), the random effects trace penalty (\code{trace_penalty}),
the number of actively estimated fixed effects (\code{p}), and the total number of random effects (\code{q}).
}
\description{
Computes the Conditional Akaike Information Criterion (cAIC) using Method 2 from
Zheng, Cadigan, and Thorson (2024).
}
\references{
Zheng, Y., Cadigan, N. G., & Thorson, J. T. (2024). A distributionally robust conditional
Akaike information criterion for random-effect models. arXiv preprint arXiv:2411.14185.
}
