% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevinSSM_utils.R
\docType{data}
\name{exampleDat}
\alias{exampleDat}
\title{Example formatted tracking data}
\description{
A \code{dataLangevin} object containing formatted movement tracks appropriate for \code{\link{fitLangevin}}..
}
