% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevinSSM_utils.R
\name{getEMF}
\alias{getEMF}
\title{Error multiplication factors}
\usage{
getEMF(
  gps = 0.1,
  emf.x = c(1, 1.54, 3.72, 13.51, 23.9, 44.22),
  emf.y = c(1, 1.29, 2.55, 14.99, 22, 32.53),
  prob = c(0, 0.05, 0.05, 0.1, 0.15, 0.25, 0.4, 0)
)
}
\arguments{
\item{gps}{A numeric value or a vector of length 2 specifying the error multiplication factor for GPS locations. If a single value is provided, it will be used for both x and y axes. Default is 0.1 (i.e. GPS errors are 10x more accurate than Argos \code{lc} 3.}

\item{emf.x}{A numeric vector of length 6 specifying the error multiplication factors for the x-axis for each location class (in order: 3, 2, 1, 0, A, B, where Z is assumed equal to B). Default values are based on the EMF values for Argos satellite telemetry data.}

\item{emf.y}{A numeric vector of length 6 specifying the error multiplication factors for the y-axis for each location class (in order: 3, 2, 1, 0, A, B, where Z is assumed equal to B). Default values are based on the EMF values for Argos satellite telemetry data.}

\item{prob}{A numeric vector of length 8 specifying the sampling probability for each location quality class (in order: G, 3, 2, 1, 0, A, B, Z). These probabilities are used internally by \code{\link{simLangevin}} and \code{\link{addMeasurementError}} to simulate observed location quality class data. Default values represent a general approximation of typical marine telemetry Argos distributions, with GPS ("G") and invalid ("Z") locations set to 0. \code{probs} must sum to 1.}
}
\value{
A data frame with columns \code{lc}, \code{emf.x}, \code{emf.y}, and \code{prob} containing the error multiplication factors and observation probabilities for each location class. The location classes included are "G" for GPS and "3", "2", "1", "0", "A", "B", and "Z" for Argos satellite telemetry data.
}
\description{
A function to generate a data frame of error multiplication factors (EMF) for different location classes, which can be used to account for measurement error for observations that lack error information but have a known location quality class. The default values are based on the EMF values for Argos satellite telemetry data, but users can specify their own EMF values for different location classes as needed. It is a modified version of the \code{emf} function from the \href{https://ianjonsen.github.io/aniMotum/}{aniMotum} package.
}
