% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskBuffer.R
\name{maskBuffer}
\alias{maskBuffer}
\title{Buffer a binary mask by a specified number of grid cells}
\usage{
maskBuffer(maskRast, bufferCells)
}
\arguments{
\item{maskRast}{A SpatRaster binary mask (1 = allowed, 0 = restricted).}

\item{bufferCells}{Integer. The number of grid cells to expand the boundary by.}
}
\value{
A SpatRaster binary mask with the padded coastline.
}
\description{
Expands the restricted areas (0s) into the allowed areas (1s) to create
a padding zone along the boundary.
}
