% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepBarrier.R
\name{prepBarrier}
\alias{prepBarrier}
\title{Prepare a spatial barrier for Langevin models}
\usage{
prepBarrier(maskRast)
}
\arguments{
\item{maskRast}{A \code{\link[terra]{SpatRaster-class}} object containing a binary mask. Values of \code{1} indicate allowed movement areas (e.g., water), and values of \code{0} indicate restricted areas (e.g., land).}
}
\value{
A \code{\link[terra]{SpatRaster-class}} object containing the calculated signed distance field. Positive values represent distance into the allowed area; negative values represent distance into the restricted area. The object is assigned a \code{"barLangevin" = TRUE} attribute.
}
\description{
Converts a binary spatial mask into a signed distance field (SDF) required for barrier constraints in the \code{langevinSSM} package. The resulting raster is tagged with a special attribute allowing downstream functions (\code{\link{fitLangevin}}, \code{\link{simLangevin}}, \code{\link{getUD}}) to automatically detect and apply the barrier penalty.
}
