% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fitLangevin.R
\name{print.fitLangevin}
\alias{print.fitLangevin}
\title{Print a fitLangevin object}
\usage{
\method{print}{fitLangevin}(x, ...)
}
\arguments{
\item{x}{A \code{fitLangevin} object returned by \code{\link{fitLangevin}}.}

\item{...}{Additional arguments passed to \code{print}.

importFrom stats printCoefmat}
}
\description{
Print a fitLangevin object
}
