% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevinSSM_utils.R
\name{print.hierLangevin}
\alias{print.hierLangevin}
\title{Print a hierLangevin object}
\usage{
\method{print}{hierLangevin}(x, ...)
}
\arguments{
\item{x}{A \code{hierLangevin} object.}

\item{...}{Ignored.}
}
\description{
Print a hierLangevin object
}
