% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevinSSM_utils.R
\name{print.regLangevin}
\alias{print.regLangevin}
\title{Print a regLangevin object}
\usage{
\method{print}{regLangevin}(x, digits = 4, header = TRUE, ...)
}
\arguments{
\item{x}{A \code{regLangevin} object returned by \code{\link{regionProb}}.}

\item{digits}{Minimal number of significant digits to print. Default: \code{4}.}

\item{header}{Logical. Indicates whether to print the main descriptive header. Default: \code{TRUE}.}

\item{...}{Additional arguments passed to \code{print}.}
}
\description{
Print a regLangevin object
}
