% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevinSSM_utils.R
\name{print.resLangevin}
\alias{print.resLangevin}
\title{Print a resLangevin object}
\usage{
\method{print}{resLangevin}(x, ...)
}
\arguments{
\item{x}{A \code{resLangevin} object returned by \code{\link{residuals.fitLangevin}}.}

\item{...}{Additional arguments passed to \code{print}.}
}
\description{
Print a resLangevin object
}
