% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevinSSM_utils.R
\name{print.summary.fitLangevin}
\alias{print.summary.fitLangevin}
\title{Print summary of a fitLangevin object}
\usage{
\method{print}{summary.fitLangevin}(x, digits = 4, signif.stars = TRUE, ...)
}
\arguments{
\item{x}{A \code{summary.fitLangevin} object.}

\item{digits}{Minimal number of significant digits, see \code{\link[stats]{printCoefmat}}.}

\item{signif.stars}{Logical. See \code{\link[stats]{printCoefmat}}.}

\item{...}{further arguments passed to \code{\link{print.default}}.}
}
\description{
Print summary of a fitLangevin object
}
