% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routeTracks.R
\name{routeTracks}
\alias{routeTracks}
\title{Reroute tracks around barriers using \code{pathroutr}}
\usage{
routeTracks(data, maskRast)
}
\arguments{
\item{data}{A formatted \code{dataLangevin} object returned by \code{\link{formatData}}.}

\item{maskRast}{A \code{\link[terra]{SpatRaster-class}} object containing a binary mask. Values of \code{1} indicate allowed movement areas (e.g., water), and values of \code{0} indicate restricted areas (e.g., land).}
}
\value{
The original \code{dataLangevin} data frame, appended with two new columns: \code{mu.x_pr} and \code{mu.y_pr}. These represent the re-routed tracks.
}
\description{
This function converts a binary barrier mask into an \code{sf} polygon and uses the \code{pathroutr} package to route observed coordinates around the barrier. It returns the original \code{dataLangevin} object with routed coordinates (\code{mu.x_pr}, \code{mu.y_pr}) appended, which can then be passed to \code{\link{fitLangevin}} as initial values for the latent locations (\code{mu}).
}
\details{
\code{\link{fitLangevin}} can sometimes struggle to push tracks around complex coastlines. \code{routeTracks} builds a visibility network graph inside the allowed areas (\code{maskRast = 1}) and bends any invalid track segments around the barrier. The resulting coordinates are appended to the dataset as \code{mu.x_pr} and \code{mu.y_pr}.
}
\examples{
if (requireNamespace("pathroutr", quietly = TRUE)) {

  library(ggplot2)

  # format the tracking data
  formatDat <- formatData(unformatDat, time.unit = "hours")

  # create a dummy mask (e.g., 1 = water, 0 = land)
  mask_rast <- exampleCovs[[1]]
  coords <- terra::crds(mask_rast)
  terra::values(mask_rast) <- ifelse(coords[, "x"] >= 1010, 1, 0)
  names(mask_rast) <- "coast_barrier"

  # route the tracks around barrier
  formatDat_routed <- routeTracks(formatDat, maskRast = mask_rast)

  barrier <- prepBarrier(mask_rast)
  plot(formatDat, spatialCovs = list(coast_barrier=barrier),
       maskRast = mask_rast)$coast_barrier +
  ggplot2::geom_path(data = formatDat_routed,
                     aes(x = mu.x_pr, y = mu.y_pr, group = id, color = as.factor(id)),
                     linewidth = 0.8) +
  ggplot2::geom_point(data = formatDat_routed,
                      aes(x = mu.x_pr, y = mu.y_pr, color = as.factor(id)),
                      shape = 16, size = 1.5) +
  ggplot2::scale_color_manual(
    name = "Track ID",
    values = c(
      "1" = "yellow",
      "2" = "lightblue",
      "3" = "orange",
      "Observed" = "lightgrey"
    )
  )
}

}
