% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCov.R
\name{simCov}
\alias{simCov}
\title{Simulate spatial covariate}
\usage{
simCov(sca = 100, irange = 0.3, sigma2 = 0.1, kappa = 0.5, M = NULL, N = NULL)
}
\arguments{
\item{sca}{Numeric value for the spatial scale of the covariate. The resulting covariate will be on a grid of size (2*sca+1) x (2*sca+1). Default: 100.}

\item{irange}{Numeric value for the spatial range of the covariate, expressed as a proportion of \code{sca}. The resulting covariate will have spatial range equal to \code{irange} * \code{sca}. Default: 0.3.}

\item{sigma2}{Numeric value for the variance of the covariate. Default: 0.1.}

\item{kappa}{Numeric value for the smoothness of the Matérn covariance function. Default: 0.5 (exponential covariance).}

\item{M}{Numeric value for the number of rows in the grid used for FFT padding. Default: NULL, in which case the optimal number of columns is dynamically calculated.}

\item{N}{Numeric value for the number of columns in the grid used for FFT padding. Default: NULL, in which case the optimal number of columns is dynamically calculated.}
}
\value{
A \code{\link[terra]{SpatRaster-class}} object containing the simulated spatial covariate.
}
\description{
Simulate a spatial covariate using a Gaussian random field with a Matérn covariance function. The resulting covariate is on a grid of size (2*sca+1) x (2*sca+1) and has spatial range equal to \code{irange} * \code{sca}.
The simulation is performed using the \code{\link[fields]{matern.image.cov}} function from the \pkg{fields} package, which uses FFT padding to efficiently simulate large spatial fields.
}
