% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevinSSM_utils.R
\docType{data}
\name{unformatDat}
\alias{unformatDat}
\title{Example unformatted tracking Data}
\description{
A data frame containing example movement tracks appropriate for \code{\link{formatData}}.
}
