% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMeasurementError.R
\name{addMeasurementError}
\alias{addMeasurementError}
\title{Add measurement error to true locations}
\usage{
addMeasurementError(
  data,
  par = NULL,
  measurementError = NULL,
  coord = c("mu.x", "mu.y")
)
}
\arguments{
\item{data}{A data frame containing the true locations (columns specified by `coord`) and optionally measurement error data (e.g., `smaj`, `smin`, `eor` for KF or `x.sd`, `y.sd` for LS/GPS).}

\item{par}{A list of parameters for the error model. For KF, this can include `psi`. For LS/GPS, this can include `tau` (as a vector of length 2) and `rho_o`. See \code{\link{fitLangevin}}.}

\item{measurementError}{A list of measurement error parameters. For KF, this should include `smaj.sd`, `smin.sd`, and optionally `eor`. For LS/GPS, this should include `x.sd` and `y.sd`. If not provided, the function will look for appropriate columns in `data`.}

\item{coord}{A character vector of length 2 specifying the column names in `data` that contain the true x and y locations (default is `c("mu.x", "mu.y")`).}
}
\value{
A data frame with observed locations (`x`, `y`). and measurement error terms added (if applicable).
}
\description{
This function adds measurement error to the true locations in the data frame, either by using provided measurement error parameters or by using existing error columns in the data. It supports both Argos Kalman Filter (KF) and Least Squares (LS)/GPS error models
}
