% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLangevin.R
\name{fitLangevin}
\alias{fitLangevin}
\title{Fit the habitat-driven Langevin diffusion}
\usage{
fitLangevin(
  data,
  model = c("underdamped", "overdamped"),
  spatialCovs,
  barrier = NULL,
  par,
  lambda = NULL,
  prior = NULL,
  map = NULL,
  coord = c("x", "y"),
  scaleFactor = 1,
  smoothGradient = FALSE,
  npoints = 4,
  curweight = 0.5,
  zetaScale = 1,
  hessian = FALSE,
  silent = FALSE,
  method = "BFGS",
  initialInner = TRUE,
  inner.control = list(maxit = 1000),
  control = list(trace = 0, iter.max = 1000, eval.max = 1000),
  polishOptim = FALSE,
  getJointPrecision = FALSE
)
}
\arguments{
\item{data}{\code{dataLangevin} object. See \code{\link{formatData}}.}

\item{model}{Character string indicating which Langevin diffusion model to fit (``underdamped'', or ``overdamped''). Default: ``underdamped''.}

\item{spatialCovs}{List of named \code{\link[terra]{SpatRaster-class}} objects containing the spatial covariates.}

\item{barrier}{Character string. The name of the barrier in \code{spatialCovs} that is represented as a signed distance field (see \code{\link{prepBarrier}}). If provided, this raster is exclusively used for the barrier penalty and is not included in the habitat selection covariates. Default: \code{NULL} (no barrier).}

\item{par}{List containing the initial values for the parameters. These can include state process parameters for the habitat selection coefficients (``beta''), the ``speed'' parameter (``sigma''), and, for the underdamped model, the friction coefficient (``gamma'').
Observation process parameters include a scaling factor to account for uncertainty in the Argos error ellipse (``psi''), a scaling factor to account for uncertainty in the x- and y-axis errors for Argos least squares or GPS observations (``tau''), and a correlation term between the x- and y-axis errors for Argos least squares or GPS observations (``rho_o''). All parameter are specified on their natural scale and are converted to working scale internally.
Any missing state process parameters are generated using \code{\link{initialValues}}. Any missing observation process parameters are fixed to their default values (``psi'' = 1, ``tau'' = 1, and ``rho_o'' = 0) via \code{map}. See Details.}

\item{lambda}{Numeric. The penalty weight for the barrier constraint. Default: \code{NULL}. If \code{NULL}, the function will attempt to automatically extract a penalty value if \code{data} is a simulated object created by \code{\link{simLangevin}}. Otherwise, it must be provided. See Details and \code{\link{suggestLambda}} for a strategy to determine the optimal penalty.}

\item{prior}{Optional 2-column data frame containing the mean (column 1) and standard deviation (column 2) for normally distributed priors on the working scale parameters. The row names must match the working scale parameter names (e.g., \code{"beta_cov1"}, \code{"log_sigma"}, \code{"log_gamma"}). Supplying the base name of a vector or matrix parameter (e.g., \code{"beta"}, \code{"mu"}, \code{"vel"}) will apply the prior to all of its elements. To target specific coordinates and time steps for the random effects, append the coordinate (\code{.x} or \code{.y}) and the row index of the observation to the base name (e.g., \code{"mu.x_1"} for the x-coordinate of the 1st observation in \code{data}, or \code{"vel.y_10"} for the y-velocity of the 10th observation in \code{data}). Parameters omitted from this data frame are assigned flat (improper) priors. Default: \code{NULL} (no priors).}

\item{map}{List defining how to optionally collect and fix parameters. See \code{\link[TMB]{MakeADFun}}.}

\item{coord}{Character vector identifying the coordinate names for the location data. Default: \code{c("x","y")}.}

\item{scaleFactor}{Internal scaling factor for the coordinates and parameters. In some cases, setting \code{scaleFactor>1} can help with optimization.}

\item{smoothGradient}{Logical indicating whether or not to smooth the gradients. See Details. Default: \code{FALSE}.}

\item{npoints}{Number of smoothing points around current cell (4 = diagonal, 8 = queen neighborhood). Ignored unless \code{smoothGradient=TRUE}.}

\item{curweight}{Smoothing weight of current cell location (\code{0 <= curweight < 1}). Ignored unless \code{smoothGradient=TRUE}.}

\item{zetaScale}{Scale factor for smooth gradient neighborhood (\code{zetaScale>1} increases and \code{zetaScale<1} decreases the neighborhood). Ignored unless \code{smoothGradient=TRUE}.}

\item{hessian}{Logical indicating whether or not to calculate the Hessian at the optimum. See \code{\link[TMB]{MakeADFun}}. Default: \code{FALSE}.}

\item{silent}{Logical indicating whether or not to disable TMB tracing information. See \code{\link[TMB]{MakeADFun}}. Default: \code{FALSE}.}

\item{method}{Outer optimization method. Default: \code{"BFGS"}.}

\item{initialInner}{Logical indicating whether or not to first perform an inner optimization for the random effects (``mu'' and/or ``vel'') before optimizing over all parameters. Default: \code{TRUE}.}

\item{inner.control}{List controlling inner optimization. See \code{\link[TMB]{MakeADFun}}.}

\item{control}{A list of control parameters for outer optimization. See \code{\link[stats]{nlminb}}.}

\item{polishOptim}{Logical indicating whether or not to perform an additional ``polishing'' optimization after the initial optimization with \code{\link[stats]{nlminb}} has completed. Default: \code{FALSE}.}

\item{getJointPrecision}{Logical indicating whether or not to return the joint precision matrix for the random effects. Default: \code{FALSE}.}
}
\value{
\code{fitLangevin} object, i.e., a list of:
\item{par}{See \code{\link[stats]{nlminb}}}
\item{objective}{See \code{\link[stats]{nlminb}}}
\item{convergence}{See \code{\link[stats]{nlminb}}}
\item{message}{See \code{\link[stats]{nlminb}}}
\item{iterations}{See \code{\link[stats]{nlminb}}}
\item{evaluations}{See \code{\link[stats]{nlminb}}}
\item{elapsedTime}{Run time of the optimization}
\item{estimates}{List containing point estimates and standard errors for the natural scale parameters (``natural''), the working scale parameters (``working''), and the random effects (``random''), where ``random'' is itself a list containing point estimates (``est'') and standard errors (``se'') for the true locations (``mu'') and/or the true velocities (``vel'').}
\item{covariance}{List containing the covariance matrices for the natural scale parameters (``natural''), the working scale parameters (``working''), and, if \code{getJointPrecision=TRUE}, the joint covariance matrix for the random effects (``random'')}.
\item{conditions}{List containing the optimization settings}
\item{signatures}{List containing lightweight fingerprints for \code{data} and \code{spatialCovs} to protect downstream functions}
\item{tmb_setup}{Blueprint for reconstructing TMB objective function}
}
\description{
Fit the underdamped or overdamped Langevin model to the location data provided, via numerical optimization of the log-likelihood
function using \code{TMB}. Location data can be temporally irregular and/or subject to measurement error.
}
\details{
\strong{Measurement Error Models:}
The \code{fitLangevin} function accommodates various types of measurement error depending on the provided measurement error data. It handles measurement error through two distinct internal models with the same underlying structure, which differ in how the observation error covariance matrix is constructed based on the available error information:
\deqn{(x,y) \sim N((\mu_x,\mu_y), \Sigma),}
where \eqn{(x.y)} are the observed locations, \eqn{(\mu_x,\mu_y)} are the true (latent) locations, and \eqn{\Sigma} is the observation error covariance matrix.
\itemize{
  \item \strong{Error Ellipse Model (Argos Kalman Filter):} This model is used when error ellipse data (\code{smaj}, \code{smin}, and \code{eor}) are present for an observation. The (optional) observation process parameter \code{psi} acts as a scaling multiplier specifically on the semi-minor axis (\code{smin}) to account for uncertainty in the error ellipse.
  For the error ellipse model, \eqn{\Sigma} is derived from the semi-major axis, semi-minor axis, and error ellipse orientation (see McClintock et al. 2015), with \code{psi} scaling the semi-minor axis.
  \item \strong{Standard Deviation Model (Argos Least Squares, GPS, Generic Locations):} This model is used when x- and y-axis standard deviations (\code{x.err} and \code{y.err}) are present for an observation. The (optional) observation parameter \code{tau} (a 2-vector) scales the \code{x.err} and \code{y.err}, respectively, to account for uncertainty in the x- and y-axis standard deviations, while the (optional) parameter \code{rho_o} accounts for correlation between the x- and y-axis standard deviations.
  For the standard deviation model, \deqn{\Sigma = \begin{bmatrix} \tau_1^2 \sigma_x^2 & \rho_o \tau_1 \tau_2 \sigma_x \sigma_y \\ \rho_o \tau_1 \tau_2 \sigma_x \sigma_y & \tau_2^2 \sigma_y^2 \end{bmatrix},}
  with \eqn{\text{tau}=(\tau_1,\tau_2)} scaling the standard deviations \eqn{(\text{x.err}=\sigma_x,\text{y.err}=\sigma_y)} and rho_o \eqn{=\rho_o} accounting for the correlation between the x- and y-axis errors.
}

\strong{Observation Process Parameters (\code{psi}, \code{tau}, \code{rho_o}):}
These parameters control the scaling and correlation of measurement errors and can be specified via \code{par} and \code{map}:
\itemize{
  \item \strong{1. Fixed at defaults:} If observation process parameters are omitted from \code{par}, the model automatically fixes them to neutral values: \code{psi = 1} (no scaling semi-minor axis), \code{tau = c(1, 1)} (no scaling of \code{x.err} or \code{y.err}), and \code{rho_o = 0} (no correlation between x- and y-axis errors).
  \item \strong{2. Estimated from data:} To estimate one or more of these parameters, include them in \code{par} with an initial starting value (e.g., \code{par = list(psi = 1.2)}).
  \item \strong{3. Fixed to custom values:} To fix a parameter to a specific value other than the default, include the custom value in \code{par} and explicitly map it to \code{NA} in \code{map} (e.g., \code{par = list(psi = 1.5), map = list(psi = factor(NA))}).
}

\strong{Gradient smoothing:} Setting \code{smoothGradient=TRUE} applies the smoothing approach of Blackwell and Matthiopoulos (2024) to the gradients of the habitat selection covariates (\code{spatialCovs}), which can help reduce attenuation bias in habitat selection coefficients when observations are obtained at a coarser time resolution than the underlying continuous-time movement process.
This approach smooths the gradient at each location by taking a weighted average of the gradients at neighboring locations, where the weights are determined by a Gaussian kernel based on the distance between locations and the specified neighborhood parameters (\code{npoints}, \code{curweight}, and \code{zetaScale}).
The neighborhood for the underdamped model is \code{zetaScale * sqrt(2*pi) * sigma}, while the neightborhood for the overdamped model is \code{zetaScale * sigma / sqrt(2)}. See Blackwell and Matthiopoulos (2024) for more details.

\strong{Barrier Constraints:}
When a \code{barrier} is specified, the model restricts animal movement using a physical penalty method for stochastic differential equations (SDEs).
The continuous signed distance function (SDF) created by \code{\link{prepBarrier}}, denoted as \eqn{d(\boldsymbol{\mu})}, is calculated as positive in the unrestricted areas (allowed space) and negative in the restricted areas.

To prevent the animal from entering the restricted zone (where \eqn{d(\boldsymbol{\mu}) \le 0}), the model applies a harmonic oscillator penalty, treating the boundary like a spring. The potential energy (\eqn{U}) of this spring is a quadratic function of the distance traveled into the restricted zone:
\deqn{U(\boldsymbol{\mu}) = \frac{1}{2} \lambda d(\boldsymbol{\mu})^2}

In Langevin dynamics, the force acting on the animal is the negative spatial gradient of the potential energy (\eqn{ -\nabla U(\boldsymbol{\mu})}). Applying the chain rule to the potential energy function yields the exact force applied during simulation and model fitting:
\deqn{\text{Force} = -\lambda d(\boldsymbol{\mu}) \nabla d(\boldsymbol{\mu})}

Because \eqn{d(\boldsymbol{\mu})} is negative in the restricted zone, the equation produces a positive force that pushes the animal in the direction of the shortest path back to the unrestricted area (\eqn{\nabla d(\boldsymbol{\mu})}).

The \code{lambda} parameter acts as the spring constant (\eqn{\lambda}); higher values create a stiffer, harder boundary. This boundary force is added directly to the spatial gradient of the habitat covariates during optimiziation, with a corresponding selection coefficient (e.g., \code{beta_barrier} if barrier name in \code{spatialCovs} is "barrier"). Because \code{lambda} is a mathematical property of the SDE rather than a biological parameter, it cannot be safely estimated via maximum likelihood estimation. It must be provided as a fixed hyperparameter (see \code{\link{suggestLambda}} for a strategy to determine the optimal penalty).
}
\examples{
# fit underdamped model with measurement error
# exampleDat included in package; see ?exampleDat for details
# exampleCovs included in package; see ?exampleCovs for details
fit <- fitLangevin(data = exampleDat,
                   spatialCovs = exampleCovs,
                   silent = TRUE)

# fit overdamped model with measurement error
fit_od <- fitLangevin(data = exampleDat,
                      model = "overdamped",
                      spatialCovs = exampleCovs,
                      silent = TRUE)

\dontrun{
# simulating with a barrier and passing the penalty to fitLangevin

par <- list(beta = c(-4, 6, 5, -0.1),
            sigma = 5,
            gamma = 0.5)

# create a dummy barrier mask (left half restricted = 0, right half allowed = 1)
coast_barrier <- exampleCovs[[1]]
terra::values(coast_barrier) <- ifelse(terra::crds(coast_barrier)[, "x"]
                                 >= mean(terra::crds(coast_barrier)[, "x"]), 1, 0)
names(coast_barrier) <- "coast_barrier"

# convert mask to SDF and add to the spatial covariates list
exampleCovs_barrier <- exampleCovs
exampleCovs_barrier$coast_barrier <- prepBarrier(coast_barrier)
exampleCovs_barrier$d2coast <- exampleCovs_barrier$coast_barrier / 100

# add a beta coefficient for d2coast to the parameter list
par_barrier <- par
par_barrier$beta <- c(par_barrier$beta, -0.2)

# simulate the data
set.seed(1,kind="Mersenne-Twister",normal.kind="Inversion")
simDat_barrier <- simLangevin(par = par_barrier,
                              nbAnimals = 3,
                              spatialCovs = exampleCovs_barrier,
                              barrier = "coast_barrier",
                              measurementError = list(smaj.sd = 1.5,
                                                      smin.sd = 0.75,
                                                      eor.lim = c(0,180)))

# Because simDat_barrier is a simLangevin object, fitLangevin will automatically
# detect and use the exact barrier penalty (lambda) that generated the data
fit_barrier <- fitLangevin(data = simDat_barrier,
                           spatialCovs = exampleCovs_barrier,
                           silent = TRUE)

plot(fit_barrier, data = simDat_barrier,
                  spatialCovs = exampleCovs_barrier,
                  maskRast = coast_barrier)
}

}
\references{
Dupont F, McClintock BT, Fischer J-O, Marcoux M, Hussey N, Auger-Methe M. 2025. Inferring resource selection and utilization distributions from irregular and error-prone animal tracking data using the habitat-driven Langevin diffusion.

McClintock BT, London JM, Camneron MF, Boveng PL. 2015. Modelling animal movement using the Argos satellite telemetry location error ellipse. Methods Ecol Evol 6:266–277. doi: 10.1111/2041-210X.12286.

Michelot T, Gloaguen P, Blackwell PG, Etienne M-P. 2019. The Langevin diffusion as a continuous-time model of animal movement and habitat selection. Methods Ecol Evol 10:1894–1907. doi: 10.1111/2041-210X.13275.

Michelot T, Hanks E. 2025. Multiscale modelling of animal movement with persistent dynamics. arXiv doi: 10.48550/arXiv.2406.15195

Blackwell PG, Matthiopoulos J. 2024. Joint inference for telemetry and spatial survey data. Ecology 105(12):e4457. doi: 10.1002/ecy.4457.
}
