% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLangevin.R
\name{fitLangevin}
\alias{fitLangevin}
\title{Fit the habitat-driven Langevin diffusion}
\usage{
fitLangevin(
  data,
  model = c("underdamped", "overdamped"),
  spatialCovs,
  par,
  map = NULL,
  coord = c("x", "y"),
  scaleFactor = 1,
  smoothGradient = FALSE,
  npoints = 4,
  curweight = 0.5,
  zetaScale = 1,
  hessian = FALSE,
  silent = FALSE,
  method = "BFGS",
  initialInner = TRUE,
  inner.control = list(maxit = 1000),
  control = list(trace = 0, iter.max = 1000, eval.max = 1000),
  getJointPrecision = FALSE
)
}
\arguments{
\item{data}{\code{dataLangevin} object. See \code{\link{formatData}}.}

\item{model}{Character string indicating which Langevin diffusion model to fit (``underdamped'', or ``overdamped''). Default: ``underdamped''.}

\item{spatialCovs}{List of named \code{\link[terra]{SpatRaster-class}} objects containing the spatial covariates.}

\item{par}{List containing the initial values for the parameters. These can include state process parameters for the habitat selection coefficients (``beta''), the ``speed'' parameter (``sigma''), and, for the underdamped model, the friction coefficient (``gamma'').
Observation process parameters include a scaling factor to account for uncertainty in the Argos error ellipse (``psi''), a scaling factor to account for uncertainty in the x- and y-axis errors for Argos least squares or GPS observations (``tau''), and a correlation term between the x- and y-axis errors for Argos least squares or GPS observations (``rho_o''). All parameter are specified on their natural scale and are converted to working scale internally.
Any missing state process parameters are generated using \code{\link{initialValues}}. Any missing observation process parameters are fixed to their default values (``psi'' = 1, ``tau'' = 1, and ``rho_o'' = 0) via \code{map}.}

\item{map}{List defining how to optionally collect and fix parameters. See \code{\link[TMB]{MakeADFun}}.}

\item{coord}{Character vector identifying the coordinate names for the location data. Default: \code{c("x","y")}.}

\item{scaleFactor}{Internal scaling factor for the coordinates and parameters. In some cases, setting \code{scaleFactor>1} can help with optimization.}

\item{smoothGradient}{Logical indicating whether or not to smooth the gradients using the approach of Blackwell and Matthiopoulos (2024). Default: \code{FALSE}.}

\item{npoints}{Number of smoothing points around current cell (4 = diagonal, 8 = queen neighborhood). Ignored unless \code{smoothGradient=TRUE}.}

\item{curweight}{Smoothing weight of current cell location (\code{0 <= curweight < 1}). Ignored unless \code{smoothGradient=TRUE}.}

\item{zetaScale}{Scale factor for smooth gradient neighborhood (\code{zetaScale>1} increases and \code{zetaScale<1} decreases the neighborhood). The neighborhood for the underdamped model is \code{zetaScale * sqrt(2*pi) * sigma}, while the neightborhood for the overdamped model is \code{zetaScale * sigma / sqrt(2)}. Ignored unless \code{smoothGradient=TRUE}.}

\item{hessian}{Logical indicating whether or not to calculate the Hessian at the optimum. See \code{\link[TMB]{MakeADFun}}. Default: \code{FALSE}.}

\item{silent}{Logical indicating whether or not to disable TMB tracing information. See \code{\link[TMB]{MakeADFun}}. Default: \code{FALSE}.}

\item{method}{Outer optimization method. Default: \code{"BFGS"}.}

\item{initialInner}{Logical indicating whether or not to first perform an inner optimization for the random effects (``mu'' and/or ``vel'') before optimizing over all parameters. Default: \code{TRUE}.}

\item{inner.control}{List controlling inner optimization. See \code{\link[TMB]{MakeADFun}}.}

\item{control}{A list of control parameters for outer optimization. See \code{\link[stats]{nlminb}}.}

\item{getJointPrecision}{Logical indicating whether or not to return the joint precision matrix for the random effects. Default: \code{FALSE}.}
}
\value{
\code{fitLangevin} object, i.e., a list of:
\item{par}{See \code{\link[stats]{nlminb}}}
\item{objective}{See \code{\link[stats]{nlminb}}}
\item{convergence}{See \code{\link[stats]{nlminb}}}
\item{message}{See \code{\link[stats]{nlminb}}}
\item{iterations}{See \code{\link[stats]{nlminb}}}
\item{evaluations}{See \code{\link[stats]{nlminb}}}
\item{elapsedTime}{Run time of the optimization}
\item{estimates}{List containing point estimates and standard error for the natural scale parameters (``natural''), the working scale parameters (``working''), and the random effects (``random''), where ``random'' is itself a list containing estimates for the true locations (``mu'') and/or the true velocities (``vel'').}
\item{conditions}{List containing the optimization settings}
}
\description{
Fit the underdamped or overdamped Langevin model to the location data provided, via numerical optimization of the log-likelihood
function using \code{TMB}. Location data can be temporally irregular and/or subject to measurement error.
}
\examples{
# fit underdamped model with measurement error
# exampleDat included in package; see ?exampleDat for details
# exampleCovs included in package; see ?exampleCovs for details
fit <- fitLangevin(exampleDat,
                   spatialCovs = exampleCovs)
}
\references{
Dupont F, McClintock BT, Fischer J-O, Marcoux M, Hussey N, Auger-Methe M. 2025. Inferring resource selection and utilization distributions from irregular and error-prone animal tracking data using the habitat-driven Langevin diffusion.

Michelot T, Gloaguen P, Blackwell PG, Etienne M-P. 2019. The Langevin diffusion as a continuous-time model of animal movement and habitat selection. Methods Ecol Evol 10:1894–1907. doi: 10.1111/2041-210X.13275.

Michelot T, Hanks E. 2025. Multiscale modelling of animal movement with persistent dynamics. arXiv doi: 10.48550/arXiv.2406.15195

Blackwell PG, Matthiopoulos J. 2024. Joint inference for telemetry and spatial survey data. Ecology 105(12):e4457. doi: 10.1002/ecy.4457.
}
