% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{Format data for \code{\link{fitLangevin}}}
\usage{
formatData(
  data,
  id = "id",
  date = "date",
  coord = c("x", "y"),
  lc = "lc",
  epar = c("smaj", "smin", "eor"),
  sderr = c("x.sd", "y.sd"),
  emf = NULL,
  time.unit = "hours",
  tz = "UTC"
)
}
\arguments{
\item{data}{A data frame or \code{\link[sf]{sf}} object containing the tracking data to be formatted. The data must contain columns for the animal ID (``id''), date/time (``date''), location quality class (``lc''; if applicable), coordinates (``coord''), and measurement error (``epar'' and/or ``sderr''; if applicable). The column names can be specified using the corresponding arguments of this function.}

\item{id}{Character string specifying the column name for the animal ID. Default: ``id''.}

\item{date}{Character string specifying the column name for the date/time, which must be of class \code{\link{DateTimeClasses}}. Default: ``date''.}

\item{coord}{Character vector of length 2 specifying the column names for the (projected) coordinates. Default: c("x", "y").}

\item{lc}{Character string specifying the column name for the location quality class. If this column is missing from the data, the function will attempt to automatically assign classes (see Details). Default: ``lc''.}

\item{epar}{Character vector of length 3 specifying the column names for the error ellipse parameters (semi-major axis, semi-minor axis, and error ellipse orientation). See details. Default: c("smaj", "smin", "eor").}

\item{sderr}{Character vector of length 2 specifying the column names for the standard deviations of the error for the x and y coordinates. Default: c("x.sd", "y.sd").}

\item{emf}{An optional data frame containing error multiplication factors (or standard deviations) for location quality classes. Must contain columns \code{lc}, \code{emf.x}, and \code{emf.y} (see \code{\link{get_emf}}). If provided, these values will be used to fill in \code{x.sd} and \code{y.sd} for observations where neither error ellipse (``epar'') nor standard deviations (``sderr'') information is provided in \code{data}. Default: \code{NULL}.}

\item{time.unit}{Character string specifying the time unit for the time steps. Default: ``hours''.}

\item{tz}{Character string specifying the time zone for the date/time column. Default: ``UTC''.}
}
\value{
A data frame of class \code{dataLangevin} containing the formatted tracking data. The data frame contains the following columns:
\item{id}{Animal ID}
\item{date}{Date/time of observation}
\item{dt}{Time step between observations (in \code{time.unit})}
\item{x}{x-coordinate of the location}
\item{y}{y-coordinate of the location}
\item{lc}{Location quality class}
\item{smaj}{Semi-major axis of the error ellipse}
\item{smin}{Semi-minor axis of the error ellipse}
\item{eor}{Error ellipse orientation (in radians)}
\item{x.sd}{Standard deviation of the x-coordinate error}
\item{y.sd}{Standard deviation of the y-coordinate error}
}
\description{
\code{formatData} takes a data frame or \code{sf} object containing tracking data and formats it for use with the \code{\link{fitLangevin}} function. The coordinates must be projected to a metric coordinate system (i.e. not longitude/latitude). It is a modified version of the \code{format_data} function from the \href{https://ianjonsen.github.io/aniMotum/}{aniMotum} package.
}
\details{
\strong{Location Classes and Measurement Errors:}
\code{formatData} is highly flexible in how it handles measurement error. Observations can have error ellipses (``epar''), standard deviations (``sderr''), missing errors that are automatically filled using \code{emf} (see \code{\link{get_emf}}), or no measurement error. Location quality classes (``lc'') are used to determine which observations have unknown errors that need to be filled using \code{emf}. These classes can include Argos Least Squares (3, 2, 1, 0, A, B, Z), GPS or Argos Kalman Filter (G), and Generic Locations (GL). The handling of measurement error for each type of location quality class is as follows:
\itemize{
  \item \strong{Argos Least Squares (3, 2, 1, 0, A, B, Z):} Typically lack explicit error parameters. If missing, these can be filled using the \code{emf} argument (see \code{\link{get_emf}}). However, if users explicitly provide error ellipses or standard deviations for any given observation in \code{data}, the user-specified values are preserved and \code{emf} filling is bypassed for this particular observation.
  \item \strong{GPS or Argos Kalman Filter (G):} Can be provided with error ellipses, standard deviations, or no errors (which will be filled by the \code{emf} table if provided).
  \item \strong{Generic Locations (GL):} Locations where standard deviations (\code{x.sd}, \code{y.sd}) are explicitly provided by the user. If \code{emf} is specified but no \code{x.sd} and \code{y.sd} are provided, these rows are not filled unless \code{emf} includes an entry for \code{GL} (otherwise an error is returned).
  \item \strong{No Measurement Error:} If the \code{emf} argument is \code{NULL} (the default), observations with \code{NA} for all \code{epar} columns and \code{NA} for all \code{sderr} columns are assumed to have no measurement error. Their existing location class is preserved. (Note: if the \code{lc} column is entirely missing from \code{data}, these observations will default to class "G").
}


\strong{Missing Location Classes:}
If the \code{lc} column is entirely missing from the input data, \code{formatData} will automatically guess the classes: rows with standard deviation data are assigned "GL", and all other rows are assigned "G".

\strong{Coordinate Systems and Units:}
For error ellipse data, the columns specified in \code{epar} should contain the semi-major axis (on the same scale as the coordinates), semi-minor axis (on the same scale as the coordinates), and error ellipse orientation (in degrees from north). \code{formatData} automatically converts the orientation from degrees to radians.
For x- and y-axis error data, the columns specified in \code{sderr} should contain the standard deviations of the errors for the x and y coordinates (on the same scale as the coordinates).
}
\examples{
# exampleDat included in package; see ?exampleDat for details
head(exampleDat)

formatDat <- formatData(exampleDat, time.unit = "mins")

\dontrun{
# exampleCovs included in package; see ?exampleCovs for details
fit <- fitLangevin(formatDat, spatialCovs = exampleCovs)
}
}
