% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUD.R
\name{getUD}
\alias{getUD}
\title{Compute Utilization Distribution}
\usage{
getUD(spatialCovs, fit, beta, log = TRUE)
}
\arguments{
\item{spatialCovs}{List of named \code{\link[terra]{SpatRaster-class}} objects containing the spatial covariates. The covariates must be on the same spatial grid and have the same spatial extent.}

\item{fit}{A \code{fitLangevin} object return by \code{\link{fitLangevin}}.}

\item{beta}{Numeric vector of habitat selection coefficients for the spatial covariates. The order of the coefficients must match the order of the covariates in \code{spatialCovs}.}

\item{log}{Logical indicating whether or not to return the log of the utilization distribution. Default: \code{TRUE}.}
}
\value{
A \code{\link[terra]{SpatRaster-class}} object containing the (log) utilization distribution. If the covariates in \code{spatialCovs} have time information (see \code{\link[terra]{time}}), the resulting time-dependent ``utilization distribution'' will also have time information, and each layer will correspond to a different time point.
}
\description{
Compute Utilization Distribution
}
\seealso{
\code{\link{plotRaster}} for plotting the utilization distribution and covariates.
}
