% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUD.R
\name{getUD}
\alias{getUD}
\title{Compute Utilization Distribution}
\usage{
getUD(
  spatialCovs,
  fit,
  beta,
  lambda = NULL,
  log = TRUE,
  nSims = 0,
  show_progress = TRUE,
  plot = TRUE,
  maskBarrier = FALSE
)
}
\arguments{
\item{spatialCovs}{List of named \code{\link[terra]{SpatRaster-class}} objects containing the spatial covariates. The covariates must be on the same spatial grid and have the same spatial extent.}

\item{fit}{A \code{fitLangevin} object return by \code{\link{fitLangevin}}.}

\item{beta}{Numeric vector of habitat selection coefficients for the spatial covariates. The order of the coefficients must match the order of the covariates in \code{spatialCovs}.}

\item{lambda}{Numeric. The penalty weight for the barrier constraint. Default: \code{NULL}. If \code{fit} is provided, this is extracted automatically.}

\item{log}{Logical indicating whether or not to return the log of the utilization distribution. Default: \code{TRUE}.}

\item{nSims}{Integer. Number of draws from the covariance matrix to use for estimating Monte Carlo uncertainty in the UD. If \code{nSims > 0}, the returned raster stack will include additional layers for simulated SE and CV. Default: \code{0}.}

\item{show_progress}{Logical. If \code{TRUE}, displays a progress bar for simulations. Default: \code{TRUE}.}

\item{plot}{Logical. Plot the resulting UD using \code{\link{plotUD}}? Default: \code{TRUE}.}

\item{maskBarrier}{Logical. If \code{TRUE}, restricted areas defined by the barrier are masked out (set to \code{NA}) before plotting the UD. This prevents the barrier penalty from compressing the color scale of the UD in the unrestricted areas. Set to \code{FALSE} to visualize the raw UD including the barrier penalty. Default: \code{FALSE}}
}
\value{
A \code{\link[terra]{SpatRaster}} object. It contains the (log) utilization distribution. It will also contain layers for Delta method standard errors (\code{UD_SE_delta}) and CVs (\code{UD_CV_delta}) if the covariance matrix is available. If \code{nSims > 0}, it adds simulated layers (\code{UD_SE_sim}, \code{UD_CV_sim}).
}
\description{
This function computes the utilization distribution (UD) for a given set of spatial covariates and habitat selection coefficients. It also estimates uncertainty in the UD using the Delta method and Monte Carlo simulations (if \code{nSims>0}). The resulting UD and uncertainty estimates are returned as \code{\link[terra]{SpatRaster-class}} objects, which can be plotted and analyzed further.
}
\examples{
# exampleCovs included in package; see ?exampleCovs for details
UD <- getUD(exampleCovs, beta = c(-4, 6, 5, -0.1) )
}
\seealso{
\code{\link{plotUD}}, \code{\link{regionProb}}.
}
