% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialValues.R
\name{initialValues}
\alias{initialValues}
\title{Generate initial parameter values}
\usage{
initialValues(
  data,
  model = c("underdamped", "overdamped"),
  par,
  spatialCovs,
  coord = c("x", "y")
)
}
\arguments{
\item{data}{A data frame of class \code{dataLangevin} containing the formatted tracking data. See \code{\link{formatData}}.}

\item{model}{Character string specifying the movement model to be fitted. Must be either "underdamped" or "overdamped". Default: "underdamped".}

\item{par}{A list of initial parameter values. The names of the list should be a subset of c("beta","sigma","gamma","mu","vel","psi","tau","rho_o"). If a parameter is not included in the list, an empirical estimate will be used as the initial value. Default: NULL.}

\item{spatialCovs}{A list of \code{\link[terra]{SpatRaster-class}} objects containing the spatial covariates to be included in the model. The order of the covariates in the list should match the order of the coefficients in \code{par$beta}.}

\item{coord}{Character vector of length 2 specifying the column names for the coordinates in the \code{data} data frame. Default: c("x", "y").}
}
\value{
A list of initial parameter values, with names corresponding to the parameters used in the model. The list will include the following parameters:
\item{beta}{Numeric vector of initial values for the coefficients of the spatial covariates. Length should match the number of spatial covariates.}
\item{sigma}{Numeric value for the initial estimate of the diffusion (or speed) parameter}
\item{gamma}{Numeric value for the initial estimate of the friction parameter (only for the underdamped model).}
\item{mu}{Numeric matrix of initial values for the true locations of the animals. Should have the same number of rows as the number of observations in the data and 2 columns for the x and y coordinates.}
\item{vel}{Numeric matrix of initial values for the true velocities of the animals (only for the underdamped model). Should have the same number of rows as the number of observations in the data and 2 columns for the x and y velocity components.}
And, if provided in \code{par}, the following observation process parameters:
\item{psi}{Numeric value for the scaling factor of the Argos KF error ellipse model.}
\item{tau}{Numeric vector of length 2 for the scaling factors of the x and y standard deviations in the LS/GPS error model.}
\item{rho_o}{Numeric value for the correlation parameter in the LS/GPS error model. Must be >=0 and <1.}
}
\description{
This function generates initial parameter values for fitting the Langevin movement model. It uses empirical estimates from the data to provide reasonable starting values for the optimization algorithm. The user can also specify initial values for any parameters, which will override the empirical estimates.
}
