% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevinSSM_utils.R
\name{langevin_methods}
\alias{langevin_methods}
\alias{logLik.fitLangevin}
\alias{coef.fitLangevin}
\alias{vcov.fitLangevin}
\alias{confint.fitLangevin}
\alias{residuals.fitLangevin}
\title{S3 Methods for Langevin model fits}
\usage{
\method{logLik}{fitLangevin}(object, ...)

\method{coef}{fitLangevin}(object, type = "natural", ...)

\method{vcov}{fitLangevin}(object, type = "natural", ...)

\method{confint}{fitLangevin}(object, parm, level = 0.95, type = "natural", ...)

\method{residuals}{fitLangevin}(object, ...)
}
\arguments{
\item{object}{A \code{fitLangevin} object.}

\item{...}{Further arguments passed to or from other methods.}

\item{type}{Character string indicating which scale to extract. Options are \code{"natural"} (default), \code{"working"}, \code{"mu"}, or \code{"vel"} (depending on the method).}

\item{parm}{A specification of which parameters are to be given confidence intervals.}

\item{level}{The confidence level required. Default: \code{0.95}.}
}
\description{
Standard S3 methods for extracting information from \code{fitLangevin} objects.
}
