% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevinSSM_utils.R
\name{langevin_methods}
\alias{langevin_methods}
\alias{print.fitLangevin}
\alias{logLik.fitLangevin}
\alias{coef.fitLangevin}
\alias{fitted.fitLangevin}
\alias{vcov.fitLangevin}
\alias{confint.fitLangevin}
\alias{summary.fitLangevin}
\title{S3 Methods for Langevin model fits}
\usage{
\method{print}{fitLangevin}(x, ...)

\method{logLik}{fitLangevin}(object, ...)

\method{coef}{fitLangevin}(object, type = c("natural", "working"), ...)

\method{fitted}{fitLangevin}(object, parm = c("mu", "vel"), ...)

\method{vcov}{fitLangevin}(object, parm = "fixed", type = c("natural", "working", "random"), ...)

\method{confint}{fitLangevin}(
  object,
  parm = "fixed",
  level = 0.95,
  type = c("natural", "working", "random"),
  ...
)

\method{summary}{fitLangevin}(object, ...)
}
\arguments{
\item{x}{A \code{fitLangevin} object returned by \code{\link{fitLangevin}}.}

\item{...}{Additional arguments passed to \code{print}.}

\item{object}{A \code{fitLangevin} object.}

\item{type}{Character string indicating which scale or type of estimates to extract. Options are \code{"natural"} (default), \code{"working"}, or \code{"random"}. Ignored for \code{fitted}.}

\item{parm}{A specification of which parameters to extract. For fixed parameters (when \code{type} is \code{"natural"} or \code{"working"}), this can be a vector of parameter names, or \code{"fixed"} (default) to return all fixed parameters. For random effects (when \code{type = "random"}, or when using \code{fitted}), this must be either \code{"mu"} (default) or \code{"vel"}.}

\item{level}{The confidence level required. Default: \code{0.95}.}
}
\value{
\itemize{
  \item \strong{\code{logLik}:} Returns an object of class \code{logLik} containing the maximized log-likelihood, with attributes for degrees of freedom (\code{df}) and number of observations (\code{nobs}).
  \item \strong{\code{coef}:} Returns a named numeric vector of parameter point estimates. If \code{type = "random"}, returns a data frame of the latent states.
  \item \strong{\code{vcov}:} Returns the variance-covariance matrix of the estimated parameters. If \code{type = "random"}, returns the sparse joint precision matrix of the random effects.
  \item \strong{\code{confint}:} For fixed parameters, returns a numeric matrix with lower and upper bounds. For random effects, returns a wide data frame containing the \code{id}, \code{date}, and the upper and lower confidence bounds for the coordinates.
  \item \strong{\code{fitted}:} Returns a data frame containing the \code{id}, \code{date}, and the estimated expected values of the latent states.
  \item \strong{\code{summary}:} Returns a \code{summary.fitLangevin} object containing convergence details and matrices of coefficients with standard errors (and Z-test p-values for habitat selection coefficients).
}
}
\description{
Standard S3 methods for extracting information from \code{fitLangevin} objects.
}
