% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRaster.R
\name{plotRaster}
\alias{plotRaster}
\title{Plot Raster}
\usage{
plotRaster(rast, legend.title = NULL)
}
\arguments{
\item{rast}{A \code{\link[terra]{SpatRaster-class}} object to plot.}

\item{legend.title}{Character string for the legend title. Default: \code{NULL}, in which case the name of the first raster layer will be used as the legend title.}
}
\value{
A ggplot object containing the raster plot.
}
\description{
Plot a \code{\link[terra]{SpatRaster-class}} object using \pkg{ggplot2} and \pkg{viridis}.
Supports multi-layer rasters by automatically creating a faceted multi-panel plot.
}
\examples{
# exampleCovs included in package; see ?exampleCovs for details
UD <- getUD(exampleCovs, beta = c(-4, 6, 5, -0.1) )
plotRaster(UD, legend.title = paste0("log(",expression(pi),")"))

}
