% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResiduals.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Plot One-Step-Ahead (OSA) Residuals for \code{fitLangevin} model}
\usage{
plotResiduals(fit, osa = NULL, tracks = NULL)
}
\arguments{
\item{fit}{A \code{fitLangevin} object returned by \code{\link{fitLangevin}}.}

\item{osa}{Data frame returned by \code{\link{getOSA}} containing the residuals. Only must be provided if \code{fit} does not already include residuals. If \code{!is.null(fit$osa)} and \code{osa} is provided, then the residuals from \code{osa} (and not \code{fit$osa}) are plotted.}

\item{tracks}{Optional. Vector of track IDs to plot separately, or \code{"all"} to plot each track in the dataset individually. If \code{NULL} (default), residuals for all tracks are aggregated into a single set of plots.}
}
\value{
List of \code{ggplot} objects (or a nested list of \code{ggplot} objects if plotting by track).
}
\description{
Generates ggplot2 diagnostic plots (Q-Q plots and ACF plots) for the OSA residuals of a fitted Langevin diffusion model. Plots are made separately for the x and y observed location residuals (which should follow a normal distribution if the model fits well), as well as for the squared Mahalanobis distance of the observed location residuals (which should follow a Chi-Square distribution with 2 degrees of freedom if the model fits well).
}
\examples{
par <- list(beta = c(-4, 6, 5, -0.1), sigma = 5, gamma = 0.5)
measurementError <- list(smaj.sd = 1.5, smin.sd = 0.75, eor = c(0,180))

set.seed(1, kind="Mersenne-Twister", normal.kind="Inversion")
# ridiculously small dataset for example purposes
smallDat <- simLangevin(par = par,
                          spatialCovs = exampleCovs,
                          nbAnimals = 3,
                          obsPerAnimal = 50,
                          measurementError = measurementError)

fit <- fitLangevin(model = "underdamped",
                   data = smallDat,
                   spatialCovs = exampleCovs,
                   silent = TRUE,
                   control = list(trace = 1),
                   calcOSA = TRUE)

plotResiduals(fit)

plotResiduals(fit, tracks = c("1", "3"))
}
