% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predLangevin.R
\name{predLangevin}
\alias{predLangevin}
\title{Predict unobserved locations from a fitted Langevin model}
\usage{
predLangevin(fit, data, spatialCovs, ...)
}
\arguments{
\item{fit}{A \code{fitLangevin} object.}

\item{data}{A \code{dataLangevin} object generated by \code{\link{formatData}}. To predict unobserved times, this dataset must have been formatted with the \code{predTimes} argument so that it contains padded \code{NA} observations at the desired prediction times.}

\item{spatialCovs}{List of named \code{\link[terra]{SpatRaster-class}} objects containing the spatial covariates. Must be the exact same list used to fit the model.}

\item{...}{Additional arguments passed to \code{\link{fitLangevin}}, e.g., \code{inner.control} to govern the TMB optimizer, \code{par} to supply custom starting values for the locations (\code{par$mu}) and/or velocities (\code{par$vel}), or \code{map} to supply fixed values for the locations (\code{map$mu}) and/or velocities (\code{map$vel}).}
}
\value{
A \code{fitLangevin} object containing the predicted trajectory. The structure is identical to a standard fitted model (with the addition of a \code{predLangevin} class), allowing it to be passed to \code{\link{plot.fitLangevin}} and other downstream functions.
}
\description{
Generates predicted locations (and velocities, for the underdamped model) along with their standard errors at unobserved times. This function works by freezing the maximum likelihood estimates of the movement and observation parameters from a previously fitted model and running an inner optimization to predict the latent locations.
}
\details{
This function provides an efficient way to interpolate animal tracks (e.g., along a regular temporal grid) after the primary model has been fitted on unpadded data. The fixed parameters are frozen and \code{TMB} optimizes over the random effects, allowing for imputation of the missing locations.
}
