% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterOverlap.R
\name{rasterOverlap}
\alias{rasterOverlap}
\title{Calculate Bhattacharyya's Affinity between two \code{\link[terra]{SpatRaster-class}} objects}
\usage{
rasterOverlap(r1, r2)
}
\arguments{
\item{r1}{A \code{\link[terra]{SpatRaster-class}} objects object. If any values are negative, \code{rasterOverlap} assumes \code{r1} is on the log scale, and the raster will be exponentiated before calculating the affinity (and a warning will be triggered).}

\item{r2}{A \code{\link[terra]{SpatRaster-class}} objects object. If any values are negative, \code{rasterOverlap} assumes \code{r2} is on the log scale, and the raster will be exponentiated before calculating the affinity (and a warning will be triggered).}
}
\value{
A numeric value between 0 (no overlap) and 1 (identical distributions).
}
\description{
This function calculates the Bhattacharyya's affinity (also known as the Bhattacharyya coefficient) between two rasters, which is a measure of similarity between two probability distributions. The rasters are first normalized to sum to 1 (to represent probability distributions), and then the affinity is calculated as the sum of the square root of the product of the two distributions across all cells.
}
