% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestLambda.R
\name{suggestLambda}
\alias{suggestLambda}
\title{Suggest a barrier penalty}
\usage{
suggestLambda(data, spatialCovs, barrier, ...)
}
\arguments{
\item{data}{A \code{dataLangevin} object containing the tracking data.}

\item{spatialCovs}{List of named \code{\link[terra]{SpatRaster-class}} objects containing the spatial covariates.}

\item{barrier}{Character string. The name of the barrier in \code{spatialCovs} that is represented as a signed distance field (see \code{\link{prepBarrier}}). If missing, it will attempt to extract the barrier name from the data attributes.}

\item{...}{Additional arguments passed to \code{\link{fitLangevin}}.}
}
\value{
A numeric value representing the suggested maximum barrier penalty (\code{lambda}).
}
\description{
Fits a baseline unconstrained model to estimate the animal's speed and uses the maximum time step
to suggest a stable upper limit for the barrier penalty (\code{lambda}). This is just a suggestion and is not necessarily the optimal value, but it can help guide users in selecting a reasonable starting point for \code{lambda} when fitting models with barriers.
}
